glassbox/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
pillow>=10.0.0
scipy>=1.11.0
opencv-python>=4.8.0
kagglehub>=0.2.0
matplotlib>=3.7.0
